Implemented Most of The Encoding Process
(The saving to binary needs to be resolved to use far fewer bytes. The pickle library inflates the
data with too much information about the object rather than only the necessary data)

(The encoding to a binary string needs to be changed to not use an ascii string that is converted
to a byte array. This is too inefficient and can be improved.)